Fix Various (#84)
* 起動時にDBを読み込む処理をリファクタリングした

* DBビュワーを少し修正

* ゲームファイルからのインポート処理を改善

* X4_ComplexCalculator.DB.X4DB をリファクタリングした

* CatFile のテストを追加

* ・メモリリークの原因のいくつかを解消
・X4 の言語ファイルでページ ID が省略された場合でも正しく解決できるようにした

* 以下の項目を編集する際、マウスホバーでフォーカスが当たるようにした
  - モジュール一覧の建造方式コンボボックス
  - 製品一覧の単価スライダー
  - 建造リソース一覧の単価スライダー

* ・NuGet パッケージの更新
・LibX4.Tests にて、テストデータ格納先パスの取得方法を修正

* LibX4.Tests にて、テストデータ格納先パスの取得方法を修正

* 一部のコメントの修正・追加

* モジュール装備インポート画面にて表示されるレコードの装備が空になる不具合の対策

* LibX4.Tests のテストデータの改行コードを LF に統一した

* 言語ファイルの内容を整理した

* LibX4 のテストデータを修正

* LibX4.Tests で使用するテストデータの見直し等

* 艦船種別が間違って抽出されているのを修正

* ・艦船のロードアウト情報が抽出されない場合があるのを修正
・DBビュワーの艦船一覧において、艦船サイズが表示されないのを修正
・DBビュワーの艦船一覧において、ゼノン/カハーク船の速度が誤っていた

* 余計な処理を削除

* 艦船一覧の速度と操舵性能にツールチップを表示するようにした

* 日光を小数指定可能にした

* DB ビュワーでフォーカスが外れてもセルが灰色にならないようにした

* 数値フィルタのテキストを左揃えにした

+++ X4_ComplexCalculator/Main/WorkArea/WorkAreaData/StationSettings/IStationSettings.cs
@@ -28,6 +28,6 @@ namespace X4_ComplexCalculator.Main.WorkArea.WorkAreaData.StationSettings
         /// <summary>
         /// 日光[%]
         /// </summary>
-        public int Sunlight { get; set; }
+        public double Sunlight { get; set; }
     }
 }
